rename: Main 컴포넌트의 파일 위치 변경
